# docs/swagger.yaml
basePath: /v2
definitions:
  responses.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: boolean
      message:
        type: string
    type: object
...
  v2controllers.BalanceResponse:
    properties:
      balance:
        type: integer
      currency:
        type: string
      unit:
        type: string
    type: object
  v2controllers.BalancesResponse:
    properties:
      balances:
        additionalProperties:
          type: integer
        type: object
    type: object
...
info:
  contact:
    name: Nostrassets
  description: Accounting wrapper for the Lightning Network providing separate accounts
    for end-users.
  license:
    name: GNU GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  title: tahub.go
  version: 0.0.1
paths:
...
  /v2/balance/:asset_id:
    get:
      consumes:
      - application/json
      description: Current user's balance in satoshi
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2controllers.BalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - OAuth2Password: []
      summary: Retrieve balance
      tags:
      - Account
  /v2/balances/all:
    get:
      consumes:
      - application/json
      description: Retrieve all user balances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2controllers.BalancesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Retrieve all balances
      tags:
      - Account
...
schemes:
- https
- http
swagger: "2.0"